Split recovery reconciliation family boundaries
